infra/requirements.txt
-e .

boto3==1.16.22
python-dotenv==0.15.0
tomlkit==0.7.0
